requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
google-genai
beautifulsoup4
pdfplumber
